Fully deprecate `TaskResult`
It's been soft-deprecated since 1.4.0.
