+++ Examples/Integration/Integration/iOS 16/IdentifiedListTestCase.swift
@@ -76,7 +76,7 @@ struct IdentifiedListView: View {
         case .incrementFirstButtonTapped:
           state.rows[id: state.rows.ids[0]]?.count += 1
           return .none
-        case .removeButtonTapped(id: let id):
+        case .removeButtonTapped(let id):
           state.rows.remove(id: id)
           return .none
         case .rows:
