+++ Sources/ComposableArchitecture/Documentation.docc/Extensions/Deprecations/EffectDeprecations.md
@@ -1,0 +1,14 @@
+# Deprecations
+
+Review unsupported reducer APIs and their replacements.
+
+## Overview
+
+Avoid using deprecated APIs in your app. Select a method to see the replacement that you should use
+instead.
+
+## Topics
+
+### Deprecated types
+
+- ``TaskResult``
