--- Sources/ComposableArchitecture/Documentation.docc/Extensions/TaskResult.md
@@ -1,23 +1,0 @@
-# ``ComposableArchitecture/TaskResult``
-
-## Topics
-
-### Representing a task result
-
-- ``success(_:)``
-- ``failure(_:)``
-
-### Converting a throwing expression
-
-- ``init(catching:)``
-
-### Accessing a result's value
-
-- ``value``
-
-### Transforming results
-
-- ``map(_:)``
-- ``flatMap(_:)``
-- ``init(_:)``
-- ``Swift/Result/init(_:)``
